requirements.txt
# For API and file processing
pyyaml==6.0.1
python-dotenv==1.0.0
tqdm==4.66.1
requests==2.31.0

# For LLM
google-generativeai==0.3.1
routeros_api

# For vector embeddings and storage
sentence-transformers==2.2.2
qdrant-client==1.7.0